+++ src/collective/exportimport/export_other.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from Acquisition import aq_base
+from collections import defaultdict
 from OFS.interfaces import IOrderedContainer
 from operator import itemgetter
 from plone import api
@@ -454,6 +455,86 @@ class ExportDiscussion(BrowserView):
         portal = api.portal.get()
         portal.ZopeFindAndApply(portal, search_sub=True, apply_func=get_discussion)
         return results
+
+
+class ExportPortlets(BrowserView):
+    def __call__(self):
+        self.title = 'Export portlets'
+        if not self.request.form.get("form.submitted", False):
+            return self.index()
+
+        all_portlets = self.all_portlets()
+        data = json.dumps(all_portlets, indent=4)
+        filename = "portlets.json"
+        self.request.response.setHeader("Content-type", "application/json")
+        self.request.response.setHeader("content-length", len(data))
+        self.request.response.setHeader(
+            "Content-Disposition", 'attachment; filename="{0}"'.format(filename)
+        )
+        return self.request.response.write(safe_bytes(data))
+
+    def all_portlets(context=None):
+        results = []
+
+        def get_portlets(obj, path):
+            uid = IUUID(obj, None)
+            if not uid:
+                return
+            portlets = export_portlets(obj)
+            if portlets:
+                results.append({"uuid": uid, "portlets": portlets})
+            return
+
+        portal = api.portal.get()
+        portal.ZopeFindAndApply(portal, search_sub=True, apply_func=get_portlets)
+        return results
+
+
+def export_portlets(obj):
+    """Serialize portlets for one content object
+    Code mostly taken from https://github.com/plone/plone.restapi/pull/669
+    """
+    from plone.app.portlets.interfaces import IPortletTypeInterface
+    from plone.portlets.interfaces import IPortletAssignmentMapping
+    from plone.portlets.interfaces import IPortletAssignmentSettings
+    from plone.portlets.interfaces import IPortletManager
+    from zope.component import getUtilitiesFor
+    from zope.component import queryMultiAdapter
+    from zope.interface import providedBy
+
+    portlets_schemata = {
+        iface: name
+        for name, iface in getUtilitiesFor(IPortletTypeInterface)
+    }
+    items = {}
+    for manager_name, manager in getUtilitiesFor(IPortletManager):
+        mapping = queryMultiAdapter((obj, manager),
+                                    IPortletAssignmentMapping)
+        if mapping is None:
+            continue
+        mapping = mapping.__of__(obj)
+        for name, assignment in mapping.items():
+            type_ = None
+            schema = None
+            for schema in providedBy(assignment).flattened():
+                type_ = portlets_schemata.get(schema, None)
+                if type_ is not None:
+                    break
+            if type_ is None:
+                continue
+            assignment = assignment.__of__(mapping)
+            settings = IPortletAssignmentSettings(assignment)
+            if manager_name not in items:
+                items[manager_name] = []
+            items[manager_name].append({
+                'type': type_,
+                'visible': settings.get('visible', True),
+                'assignment': {
+                    name: getattr(assignment, name, None)
+                    for name in schema.names()
+                },
+            })
+    return items
 
 
 def safe_bytes(value, encoding="utf-8"):